requirements.txt
asgiref==3.6.0
dj-database-url==2.0.0
Django==4.2.1
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
dnspython==2.3.0
pymongo==4.3.3
requests==2.31.0

